ci: isolate object ordering
